fix cppcheck warnings

+++ src/process_iterator_apple.c
@@ -117,7 +117,7 @@ static int get_process_pti(pid_t pid, struct proc_taskallinfo *ti)
 	}
 	else if (bytes < (int)sizeof(ti))
 	{
-		fprintf(stderr, "proc_pidinfo: too few bytes; expected %ld, got %d\n", sizeof(ti), bytes);
+		fprintf(stderr, "proc_pidinfo: too few bytes; expected %lu, got %d\n", (unsigned long)sizeof(ti), bytes);
 		return -1;
 	}
 	return 0;
